requirements.txt
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text processing and NLP
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0

# Visualization
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
wordcloud>=1.9.0

# Additional dependencies
Pillow>=10.0.0

# Google Cloud APIs for competitor semantic analysis
google-cloud-language>=2.11.0
google-cloud-aiplatform>=1.38.0
requests>=2.31.0
